requirements.txt
Flask==2.1.1
Flask-WTF==1.0.0
email-validator==1.2.1
mysql-connector-python==8.0.32
Werkzeug==2.0.3  # Specify an appropriate version
